requirements.txt
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.0.1
pandas==2.1.3
pdf2image==1.16.3
openpyxl==3.1.2
pyinstaller==6.2.0